Created By: @王同贵
Accepted By: @王同贵
URL: https://digit-force.coding.net/p/ai-platform/d/solution/git/merge/1049

+++ solution-application/src/main/java/com/digitforce/aip/facade/SceneCmdFacadeImpl.java
@@ -11,6 +11,7 @@ import com.digitforce.aip.enums.SceneStatusEnum;
 import com.digitforce.aip.service.ISceneService;
 import com.digitforce.aip.service.ISceneVersionService;
 import com.digitforce.framework.api.dto.Result;
+import com.digitforce.framework.api.exception.BizException;
 import com.digitforce.framework.context.TenantContext;
 import com.digitforce.framework.tool.ConvertTool;
 import org.springframework.transaction.annotation.Transactional;
@@ -44,6 +45,13 @@ public class SceneCmdFacadeImpl implements SceneCmdFacade {
 
     @Override
     public Result delete(SceneDeleteCmd sceneDeleteCmd) {
+        Scene scene = sceneService.getById(sceneDeleteCmd.getId());
+        if (scene == null) {
+            throw new BizException("场景不存在");
+        }
+        if (scene.getStatus() == SceneStatusEnum.ONLINE) {
+            throw new BizException("场景已上线，不能删除");
+        }
         sceneService.removeById(sceneDeleteCmd.getId());
         return Result.success();
     }
@@ -59,7 +67,13 @@ public class SceneCmdFacadeImpl implements SceneCmdFacade {
 
     @Override
     public Result unPublish(SceneStatusCmd sceneStatusCmd) {
-        Scene scene = ConvertTool.convert(sceneStatusCmd, Scene.class);
+        Scene scene = sceneService.getById(sceneStatusCmd.getId());
+        if (scene == null) {
+            throw new BizException("场景不存在");
+        }
+        if (scene.getStatus() == SceneStatusEnum.OFFLINE) {
+            throw new BizException("场景已下线，不能重复下线");
+        }
         scene.setStatus(SceneStatusEnum.OFFLINE);
         scene.setUpdateUser(TenantContext.tenant().getUserAccount());
         sceneService.updateById(scene);
@@ -67,10 +81,9 @@ public class SceneCmdFacadeImpl implements SceneCmdFacade {
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public Result modify(SceneModifyCmd sceneModifyCmd) {
-        Scene scene = ConvertTool.convert(sceneModifyCmd, Scene.class);
-        scene.setUpdateUser(TenantContext.tenant().getUserAccount());
-        sceneService.updateById(scene);
+        sceneService.updateScene(sceneModifyCmd);
         return Result.success();
     }
 }
